Split major classes and functionality into separate files. Have tested none of it.
